Move entity state tracker into its own class.
